requirements.txt
python-telegram-bot>=21.0,<22.0
python-dotenv>=1.0.0,<2.0.0
python-i18n[YAML]>=0.3.9,<0.4.0
openai>=1.0.0,<2.0.0
yt-dlp>=2025.1.15
pytest>=8.0.0,<9.0.0
